Revert "feat: scan snippets (#33)"
This reverts commit 3059aa724d51b41a34f0409aac1f305dc4ec07ad.
