+++ state.go
@@ -23,6 +23,7 @@ func (s State) Save() error {
 	}
 	defer fi.Close()
 	return json.NewEncoder(fi).Encode(s)
+
 }
 
 // defaultState returns the default state path
